SSCS-4605 Update hearing type when relisting hearing
We want to modify the hearing type when re relist an appeal from cor to
oral. This was hard before as we would of had to recreate the entire
SscsDocument. Added toBuilder=true to all the objects this allows us to
create a builder from the current state of an object.

Added new EventType UPDATE_HEARING_TYPE for this event.